refactor(orders): extract reversal off-ramps into module
Move createOrderRefund + cancelOrder and their private refund helpers
(buildRefundItems, getOrderLineRefundCaps, roundCurrencyUnit) out of the
1106-line order-admin.service.ts into a new order-reversal.service.ts, peer
of order-pickup.service.ts. order-admin.service.ts drops to 802 lines (-27%).

Pure relocation, no behavior change: same validations, error messages,
transactions and permission gates. The planned applyTerminalExit scaffold was
rejected -- the status-machine seams (transitionOrderService, applyRefundTransition)
and the permissions module already own the flip/log/recalc and payment-status
gating it would have wrapped.

Route imports repointed; docs/architecture-deepening.md section 6 marked done.
